Set proper endpoint in `run-jupyter-lab.sh`
I had a problem where I could not run `knic-jupyter` properly without
manually overriding the endpoint. The solution turned out to be removing
a quoted empty string after the `-i` parameter of `sed` because it was always
erroring. I'm not entirely sure what was going wrong, but removing the
string stopped the errors, and the script now works for me.