Specify Dockerfile syntax version
According to the Dockerfile reference:[^1]

> Most users will want to set this parser directive to
> `docker/dockerfile:1`, which causes BuildKit to pull the latest stable
> version of the Dockerfile syntax before the build.

While this is not as stable as pinning, the syntax follows SemVer, so
pinning the major version should be sufficient. Even if there is a
breakage, it's better to use the latest version for everyone than to use
whatever happens to be installed with Docker.

[^1]: https://docs.docker.com/reference/dockerfile/#syntax